+++ includes/Modules/Registry.php
@@ -42,6 +42,7 @@ final class Registry {
 		$modules = array(
 			new \QuickEventsManager\Events\EventsModule(),
 			new \QuickEventsManager\Registration\RegistrationModule(),
+			new \QuickEventsManager\Venues\VenuesModule(),
 		);
 
 		/**